backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "proposal-generator"
version = "0.1.0"
description = "クラウドソーシング案件向け提案文自動生成CLIツール"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
keywords = ["cli", "proposal", "crowdsourcing", "lancers", "crowdworks", "gemini"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

dependencies = [
    "playwright>=1.40",
    "google-generativeai>=0.3",
    "httpx>=0.25",
    "typer>=0.9",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pyperclip>=1.8",
    "rich>=13.0",
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "supabase>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "black>=23.0",
    "isort>=5.12",
    "ruff>=0.1",
    "mypy>=1.5",
    "types-PyYAML>=6.0",
    "respx>=0.20",
    "freezegun>=1.2",
]

[project.scripts]
proposal-gen = "src.main:app"

[project.urls]
Homepage = "https://github.com/your-username/proposal-generator"
Documentation = "https://github.com/your-username/proposal-generator/tree/main/docs"
Repository = "https://github.com/your-username/proposal-generator"
Issues = "https://github.com/your-username/proposal-generator/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.venv
    | __pycache__
    | \.mypy_cache
    | \.pytest_cache
    | dist
    | build
    | docs/research
)/
'''

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["src"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
skip = [".venv", "__pycache__", "docs/research"]

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long (black handles this)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict
]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    ".mypy_cache",
    ".pytest_cache",
    "dist",
    "build",
    "docs/research",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = [
    ".venv",
    "__pycache__",
    "dist",
    "build",
    "docs/research",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow",
    "e2e: marks tests as end-to-end",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
